+++ src/design/plone/contenttypes/behaviors/trasparenza.py
@@ -274,7 +274,6 @@ class ITrasparenza(model.Schema):
         RelatedItemsFieldWidget,
         vocabulary="plone.app.vocabularies.Catalog",
         pattern_options={
-            "maximumSelectionSize": 10,
             "selectableTypes": ["Persona"],
         },
     )
@@ -283,7 +282,6 @@ class ITrasparenza(model.Schema):
         RelatedItemsFieldWidget,
         vocabulary="plone.app.vocabularies.Catalog",
         pattern_options={
-            "maximumSelectionSize": 10,
             "selectableTypes": ["Persona"],
         },
     )