+++ src/design/plone/contenttypes/interfaces/pagina_argomento.py
@@ -56,7 +56,6 @@ class IPaginaArgomento(model.Schema, IDesignPloneContentType):
         RelatedItemsFieldWidget,
         vocabulary="plone.app.vocabularies.Catalog",
         pattern_options={
-            "maximumSelectionSize": 10,
             "selectableTypes": ["UnitaOrganizzativa"],
         },
     )