+++ src/design/plone/contenttypes/interfaces/punto_di_contatto.py
@@ -78,7 +78,6 @@ class IPuntoDiContatto(model.Schema, IDesignPloneContentType):
         RelatedItemsFieldWidget,
         vocabulary="plone.app.vocabularies.Catalog",
         pattern_options={
-            "maximumSelectionSize": 10,
             "selectableTypes": ["Persona"],
         },
     )